split CheckstyleValidatorTest into several test classes
