+++ qulice-checkstyle/src/test/java/com/qulice/checkstyle/test/extensions/CheckstyleValidateRunner.java
@@ -1,0 +1,63 @@
+/*
+ * Copyright (c) 2011-2024 Qulice.com
+ *
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met: 1) Redistributions of source code must retain the above
+ * copyright notice, this list of conditions and the following
+ * disclaimer. 2) Redistributions in binary form must reproduce the above
+ * copyright notice, this list of conditions and the following
+ * disclaimer in the documentation and/or other materials provided
+ * with the distribution. 3) Neither the name of the Qulice.com nor
+ * the names of its contributors may be used to endorse or promote
+ * products derived from this software without specific prior written
+ * permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT
+ * NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
+ * FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL
+ * THE COPYRIGHT HOLDER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT,
+ * INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR
+ * SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT,
+ * STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE)
+ * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED
+ * OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package com.qulice.checkstyle.test.extensions;
+
+import com.qulice.spi.Violation;
+import java.io.IOException;
+import java.util.Collection;
+
+/**
+ * Runs a checkstyle to validate the specified file.
+ *
+ * @since 0.1
+ */
+public interface CheckstyleValidateRunner {
+
+    /**
+     * Validates that checkstyle reported given violation.
+     * @param file File to check.
+     * @param result Expected validation result.
+     * @param message Message to match
+     * @throws Exception In case of error
+     */
+    void validate(String file, boolean result,
+        String message) throws Exception;
+
+    /**
+     * Returns string with Checkstyle validation results.
+     * @param file File to check.
+     * @param passes Whether validation is expected to pass.
+     * @return String containing validation results in textual form.
+     * @throws IOException In case of error
+     */
+    Collection<Violation> runValidation(String file,
+        boolean passes) throws IOException;
+}
